Add support for ELBO object as loss arg to SVI
