+++ pyro/infer/__init__.py
@@ -6,16 +6,9 @@ from pyro.infer.enum import config_enumerate
 from pyro.infer.importance import Importance
 from pyro.infer.search import Search
 from pyro.infer.svi import SVI
+from pyro.infer.trace_elbo import Trace_ELBO
+from pyro.infer.traceenum_elbo import TraceEnum_ELBO
+from pyro.infer.tracegraph_elbo import TraceGraph_ELBO
+from pyro.infer.util import enable_validation, is_validation_enabled
 
 # flake8: noqa
-
-_VALIDATION_ENABLED = False
-
-
-def enable_validation(is_validate):
-    global _VALIDATION_ENABLED
-    _VALIDATION_ENABLED = is_validate
-
-
-def is_validation_enabled():
-    return _VALIDATION_ENABLED
